添加git-credentials 【注意】 已经将含有密码的文件添加到gitignore

+++ scripts/backup.sh
@@ -9,7 +9,7 @@ sudo_cp_if_file_exists(){
 }
 
 cp_if_file_exists(){
-    if [ -f "$1" ];
+    if [ -f "$1" -a -e $2 ];
     then
         cp "$1" "$2"
     fi
@@ -26,9 +26,14 @@ local_backup(){
 	# tmux
 	cp ~/.tmux.conf tmux
 
-	# git 备份之前关闭了代理
+	# git 
+	# 备份之前关闭了代理
 	source ss/proxy.zsh
-	proxy stop && cp_if_file_exists ~/.gitconfig git && cp_if_file_exists ~/.gitmessage git && proxy start
+	proxy stop 
+	cp_if_file_exists ~/.gitconfig git
+	cp_if_file_exists ~/.gitmessage git
+	cp_if_file_exists ~/.git-credentials git
+	proxy start
 
 	# ssh
 	cp ~/.ssh/config ssh
